+++ CHANGES.rst
@@ -13,6 +13,7 @@ Unreleased
 - :meth:`.revert` to revert a WikiPage to a specified revision.
 - :meth:`.Inbox.mark_all_read` to mark all messages as read with one API call.
 - :meth:`~.InboxableMixin.unblock_subreddit` to unblock a subreddit.
+- :meth:`.update_crowd_control_level` to update the crowd control level of a post.
 
 7.3.0 (2021/06/17)
 ------------------